Add text lines to understand git commit command
I did this becuase I was trying to understand how to add a commit message that had both a subject and a body. I discoverd that using git commit without the -m option willl open my text editor and allow me to type a mulitline message. It appears to be wrapping each line at a certain number of characters and I'm hoping it is 72 characters.